:recycle: Split travel form creation/editing and manage components apart.
At this point this just means duplicating the components. In the future this should me simplifying and customizing each component.

Its generally a good idea to break components up by view or action type as this scales better and is more flexible. Of course, its only a good idea if all shared logic is broken out into shared components.
It also makes completing user stories related to a given feature easier to implement as you only have to deal with a making changes to a specific view.