:sparkles: Support pre approved travel request with requestor name.
This doesn't save any differently in the database, but the selector is now prettier.

An alternative would be to load the preapprovedTravelers table directly, but that would restrict the list to only traveler data.

BY THE POWER OF FRONT-END TRANSOFMRS: this is what you can do if your back-end gives you data in a standard format.