:construction: Beging rebuild with insta-form create, redirect to edit.
Taking advantage of the "draft" state to create a form instance immediately on request, and redirecting the user to an "edit" page for the newly created form.

This lets the user save the form at any point, and lets us calculate estimates without forcing us to put that in the form creation pipeline.