:recycle: swap more forms references to use sequelize model.
Why? The table is now called "travel_authorizations", and this makes the code more flexible.
I'll later rename the model to match the table name.

Note that I didn't migrate the "modify" query to use Sequelize as its complicated and the risk of breaking it is too high to change it.