:beetle: Fix policy scoping return amount.
Previous, the scope was applied second so the record count would be off by the number of records removed. This could also create the scenario where there were no records returned on the first page of data, but it was reported as full.

Now the scoping happens first, and the count is correct.
The downside is that the policy scope is now much more complex.

Note that without typing the `where` separately, the "scope" function rejects because of type conflicts.
This appears to be a bug in the Sequelize scope type definition?