:beetle: Fix read my travel authorization approvals pre-approved travel authorizations display.
The previous code was using travel requests, which doesn't correctly map to anything in the data model in relation to a travel authorizations.

Now I'm showing the user's name if they have special permission, or just the generic approval info if they don't. This assumes, very unsafely, that the fullname is the same across both objects.
Really, I should be using a user id, but the data model doesn't support that.