:recycle: Move per-diem seeds to a csv and add in "Incidentals".
Previously the seeding was using "Maximum Daily", but this doesn't accurately reflect the data  usage, where "Incidentals" would often need to be computed from max daily - breakfast/lunch/dinner.
As that's not very convenient, I'm making use of the provided data and have incidentals as a separate column, and "maximum daily" as a computed value.