:construction: Begin work on adding correcting line for non-travel status days.
This needs a bunch more work, and I'm not confident this is the best way to add a correcting entry.
There are a bunch of edge cases that need to be handle, like limiting the number of non-travel-status days to the maximum number of days in the trip, and stuff like that.