:recycle: Drop logging of travel authorization deletion.
Why? It's not clear how to model this behavior in the current system.

In the future it might make sense to switch to "paranoid" mode, i.e. never actually delete anything but instead set the "deleted_at" timestamp on the travel authorization instance.

Alternatively we could use the "paper trail" pattern via a package such as https://www.npmjs.com/package/sequelize-paper-trail. This pattern stores all changes against a given model as a json blob.