:cherry_blossom: Add better reactivity to expense page dependencies.
I'm starting to require reactivity between cousins (i.e deeply nested siblings).
One solution to this would be to use Vuex stores, and eat the added complexity.
And alternative would be a global event bus; maybe through a "use" file.
A third option would be to customize the "use" stores and provide a way to listen to events on a per-use file level.

e.g.

const { expenses, on } = useExpenses()

on('global:changed', () => {
   // some action
})

where "on" function would be shared between