:beetle: Fix logic for multi-destination trip prefilling.
Previously the final destination date/time information was being injected into the last stop, but this isn't editable via the UI, as the "Stop" concept doesn't encapsulate the UI logic very well.

This was then messing with the sort order of the travelAuthorization.stops property and putting the last stop as the first or second stop.

This caused the travelAuthorization.buildTravelSegmentsFromStops method to crash as the stop was malformed; it was missing a transport method.