:beetle: Move defineExpose to the end of the setup.
Why? When exposing functions exposed through use files, the defineExpose must come after the use reference.

This is bit confusing and relates to scopes.
When a function is defined via `function () {...}` it gets a higher scope and can be reference before declaration, whereas, when defined via const or from another source, it must be defined before being used.

As such, I've decided to always put defineExpose as the bottom of the setup function.