:arrow_up: Upgrade Sequelize to fixing findAll date serialization bug.
Bug in question is:
When you load a model with .findOne, it converts all date type fields to Date objects, and serializers via toJSON into ISO standard format.
When you load a model with .findAll, it does not convert date type fields to Date objects, but instead uses the raw database SQL string. When this gets serialized via toJSON, it remains a string, in SQL formatted time.

Did not work.
Code works in ts-node console, but not in app.