:ok_hand: Enforce UTC for rental car times during display.
Why? Because the time is treated as UTC when saved
i.e.
rentalCar.value.pickUpDate = pickUpDate.value + "T" + pickUpTime.value + ":00.000Z"

This is an oversight in the original code.
The end user should be made aware that times are in UTC,
or the time picker should allow picking the the time zone,
or the time should be local to the pickup/drop of location.