:recycle: Make flight time radio group a component.
Why? Its complicated enough that the code reads better with it as a component.
And now it will always works the same for the create and edit dialogs.

Special v-radio group where label is above row of radio buttons.
This is different than v-radio-group "row" layout where label is to the left of radio buttons.