:recycle: Make paranoid mode the default for all models.
Why? It helps track creation/destruction rates for data.
Also makes it much easier to undo stuff.

See https://sequelize.org/docs/v6/core-concepts/paranoid/
Paranoid mode adds a `deleted_at` column to the table.
When a record is destroyed, it's `deleted_at` is set to the current time.
When a record is restored, it's `deleted_at` is set to `null`.