:japanese_castle: Refresh tabs when state changes in child page.
After triggering a "submit" action in the TravelDeskTravelRequestEditCard,
we redirect from the MyTravelRequestsRequestEditPage to the MyTravelRequestsRequestReadPage.

However, until now, the tab path would still point to the MyTravelRequestsRequestEditPage.

Now we are refreshing the tab, to trigger a re-computation of the tab path.

I'm not sure this is the best way, but it seems the least bad?