:art: Standardize model export pattern.
Why? Have a single pattern for the model export is simpler and easier to maintain.
Yes, it might take more characters, as some model could be exported directly, but this way we need only make a single change when adding a relationship to a model. And its simpler to only export one way.