:cherry_blossom: Update travel desk export file name.
Why? use standard "type, name, date" format for file name.
Sadly, the library we use for export mangles the file name, and replaces spaces with underscores.
It isn't worth upgrading the library at this time.
In the future, we will move this code to the back-end, and use a completely different library.