:broom: Remove unused legacy store setters.
Newer code pattern is to use local state instead of global state.
If this new pattern results in many api requests, the api calls can be debounced, or back-end caching can be used instead.

The pattern of using local state for components makes components more re-usable, and easier to debug.