:bug: Fix travel authorization pre-approval profile filter/scope generation.
Previously was only applying the last scope sent due to the lack of the feature listed in
https://github.com/sequelize/sequelize/issues/17337.

Newer patterns for the base controller make use of a scope prebuild that avoid this issue.
Also update the pre-approval profile policy as it is required to make use of the newer scope pattern.